+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - Fix browser compatibility in Navbar component
 - Fix preventClick function on touch devices
+- Fix LightboxPanel props
 
 ## 3.0.0 beta 40 (February 8, 2018)